requirements.txt

fastapi==0.99.1
uvicorn==0.22.0
python-dotenv==1.0.0
twilio==8.4.0
pydantic==1.10.8
python-multipart==0.0.6
